requirements.txt
asgiref==3.7.2
camelot-py==0.11.0
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
distro==1.9.0
Django==3.2
et-xmlfile==1.1.0
ghostscript==0.7
JPype1==1.5.0
numpy==1.26.3
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
pdfminer.six==20231228
pycparser==2.21
pypdf==3.17.4
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tabula-py==2.9.0
tabulate==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
